+++ smart-contracts/master-pyteal/box_storage/grocery_app/utils.py
@@ -1,0 +1,54 @@
+from algosdk.transaction import *
+from algosdk.v2client.algod import *
+
+def create_app(
+    client,
+    sender,
+    private_key,
+    approval_program,
+    clear_program,
+    global_schema,
+    local_schema,
+):
+
+    # declare on_complete as NoOp
+    on_complete = transaction.OnComplete.NoOpOC.real
+
+    # get node suggested parameters
+    params = client.suggested_params()
+
+    # create unsigned transaction
+    txn = transaction.ApplicationCreateTxn(
+        sender,
+        params,
+        on_complete,
+        approval_program,
+        clear_program,
+        global_schema,
+        local_schema,
+    )
+
+    # sign transaction
+    signed_txn = txn.sign(private_key)
+    tx_id = signed_txn.transaction.get_txid()
+
+    # send transaction
+    client.send_transactions([signed_txn])
+
+    # await confirmation
+
+    confirmed_txn = transaction.wait_for_confirmation(client, tx_id, 4)
+    print("TXID: ", tx_id)
+    print("Result confirmed in round: {}".format(confirmed_txn["confirmed-round"]))
+
+    # display results
+    transaction_response = client.pending_transaction_info(tx_id)
+    app_id = transaction_response["application-index"]
+    print("Created new app-id: ", app_id)
+
+    return app_id
+
+
+def compile_program(client, source_code):
+    compile_response = client.compile(source_code)
+    return base64.b64decode(compile_response["result"])
